requirements.txt
Django==5.0.2
gunicorn==21.2.0
djangorestframework==3.14.0
ipadic==1.0.0
mecab==0.996.3
mecab-python3==1.0.8
nltk==3.8.1